+++ src/main/java/kr/co/itcen/bookmall/dao/test/BookDaoTest.java
@@ -17,20 +17,20 @@ public class BookDaoTest {
 		
 		BookVo vo1 = new BookVo();
 		vo1.setTitle("소설책1");
-		vo1.setPrice((long)10000);
+		vo1.setPrice(10000);
 		vo1.setCategory_no((long)1);
 		dao.insert(vo1);
 		
 		BookVo vo2 = new BookVo();
 		vo2.setTitle("인문책2");
 		vo2.setCategory_no((long)2);
-		vo2.setPrice((long)15000);
+		vo2.setPrice(15000);
 		dao.insert(vo2);
 		
 		BookVo vo3 = new BookVo();
 		vo3.setTitle("사회책3");
 		vo3.setCategory_no((long)3);
-		vo3.setPrice((long)5000);
+		vo3.setPrice(5000);
 		dao.insert(vo3);
 	}
